Switch parsing to Long
This avoids issues with potential number clashes with ID_UNSET.
Also fixes some further parsing where adaptation set ids are referenced
from manifest properties.

+++ libraries/exoplayer_dash/src/main/java/androidx/media3/exoplayer/dash/manifest/AdaptationSet.java
@@ -25,17 +25,13 @@ import java.util.List;
 public class AdaptationSet {
 
   /** Value of {@link #id} indicating no value is set.= */
-  public static final int ID_UNSET = -1;
+  public static final long ID_UNSET = -1;
 
   /**
-   * An identifier for the adaptation set that's unique in the scope of its containing period, or
-   * {@link #ID_UNSET} if not specified.
-   *
-   * <p>The identifier has the same value as defined in the manifest. However. the manifest value is
-   * an unsigned integer and values exceeding {@link Integer#MAX_VALUE} will be represented as
-   * negative numbers.
+   * A non-negative identifier for the adaptation set that's unique in the scope of its containing
+   * period, or {@link #ID_UNSET} if not specified.
    */
-  public final int id;
+  public final long id;
 
   /** The {@link C.TrackType track type} of the adaptation set. */
   public final @C.TrackType int type;
@@ -62,7 +58,7 @@ public class AdaptationSet {
    * @param supplementalProperties Supplemental properties in the adaptation set.
    */
   public AdaptationSet(
-      int id,
+      long id,
       @C.TrackType int type,
       List<Representation> representations,
       List<Descriptor> accessibilityDescriptors,
